pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "lightgbm>=4.6.0",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.45.0",
]
